Initial upload
This is a simple python script to issue voice commands to the seamoth vehicle in Subnautica running under Windows 10.  This is only a proof-of-principle code and a starting point for more advanced functionality to be added later.  The code requires the python modules SpeechRecognition and PyAudio to be installed, and it also uses the keys module taken from Sentdex's cyberpython2077 project (https://github.com/Sentdex/cyberpython2077) which was the original inspiration for this project.